fix(cosh-ng): [shell] untracked handoff fallback
When a preexec marker is lost with the guard still defeated (deep
alias chains, unknown shells), an emitted shell handoff stayed
pending forever and blocked the Agent run and input routing.

Add a prompt-boundary fallback: ShellHandoffState records the event
index at emission, and close_untracked_shell_handoffs closes the
front pending handoff once the first relevant boundary after that
index is a ShellReady - a CommandStarted seen first defers to the
tracked block path, and events after an already-reached boundary
never veto it (the event list is cumulative). Closure emits degraded
evidence (completed_untracked, exit code -1, no output), marks the
provider tool completed, and feeds the existing continuation
recovery; the provider notice states the result must not be treated
as a command outcome.

Organization: the handoff row builders (tracked and untracked) live
in the new activity/shell_handoff.rs owner file with runtime.rs as a
facade; the cross-owner status token SHELL_HANDOFF_UNTRACKED_STATUS
and the ShellHandoffRequest contract move to types/shell_handoff.rs;
the ui status label matches the shared constant; i18n gains the
untracked status label in en/zh.
